gitserver/server: Move funcs & types to common package avoid cyclic deps
