+++ cmd/gitserver/server/vcs_syncer_git.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"os/exec"
 
@@ -13,22 +12,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/wrexec"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
-
-// GitCommandError is an error of a failed Git command.
-type GitCommandError struct {
-	// Err is the original error produced by the git command that failed.
-	Err error
-	// Output is the std error output of the command that failed.
-	Output string
-}
-
-func (e *GitCommandError) Error() string {
-	return fmt.Sprintf("%s - output: %q", e.Err, e.Output)
-}
-
-func (e *GitCommandError) Unwrap() error {
-	return e.Err
-}
 
 // GitRepoSyncer is a syncer for Git repositories.
 type GitRepoSyncer struct{}
@@ -51,13 +34,13 @@ func (s *GitRepoSyncer) IsCloneable(ctx context.Context, remoteURL *vcs.URL) err
 	defer cancel()
 
 	cmd := exec.CommandContext(ctx, "git", args...)
-	out, err := runWith(ctx, wrexec.Wrap(ctx, log.NoOp(), cmd), true, nil)
+	out, err := common.RunWith(ctx, wrexec.Wrap(ctx, log.NoOp(), cmd), true, nil)
 	if err != nil {
 		if ctxerr := ctx.Err(); ctxerr != nil {
 			err = ctxerr
 		}
 		if len(out) > 0 {
-			err = &GitCommandError{Err: err, Output: string(out)}
+			err = &common.GitCommandError{Err: err, Output: string(out)}
 		}
 		return err
 	}
@@ -73,7 +56,7 @@ func (s *GitRepoSyncer) CloneCommand(ctx context.Context, remoteURL *vcs.URL, tm
 	cmd = exec.CommandContext(ctx, "git", "init", "--bare", ".")
 	cmd.Dir = tmpPath
 	if err := cmd.Run(); err != nil {
-		return nil, errors.Wrapf(&GitCommandError{Err: err}, "clone setup failed")
+		return nil, errors.Wrapf(&common.GitCommandError{Err: err}, "clone setup failed")
 	}
 
 	cmd, _ = s.fetchCommand(ctx, remoteURL)
@@ -85,8 +68,8 @@ func (s *GitRepoSyncer) CloneCommand(ctx context.Context, remoteURL *vcs.URL, tm
 func (s *GitRepoSyncer) Fetch(ctx context.Context, remoteURL *vcs.URL, dir common.GitDir, revspec string) ([]byte, error) {
 	cmd, configRemoteOpts := s.fetchCommand(ctx, remoteURL)
 	dir.Set(cmd)
-	if output, err := runWith(ctx, wrexec.Wrap(ctx, log.NoOp(), cmd), configRemoteOpts, nil); err != nil {
-		return nil, &GitCommandError{Err: err, Output: newURLRedactor(remoteURL).redact(string(output))}
+	if output, err := common.RunWith(ctx, wrexec.Wrap(ctx, log.NoOp(), cmd), configRemoteOpts, nil); err != nil {
+		return nil, &common.GitCommandError{Err: err, Output: newURLRedactor(remoteURL).redact(string(output))}
 	}
 	return nil, nil
 }
